View Demo 2
Major Changes:
Revamped entire page to allow for proper resizing (before a slightly smaller screen would've affected the view by a lot)
New colour scheme & logo to fit colour scheme
Restructure of all classes in index.jsp. Now its a lot more easier to sort out everything.
Unneccesary attributes removed
movies.jspf was added. This can serve as either a placeholder or as a base for someone else working on the list for movies.

Minor Changes:
'Coming Soon' removed
Only the price category is in refine. This is a placeholder